WIP: Added DALI

+++ image_classification/tensorflow/official/resnet/imagenet_main.py
@@ -387,14 +387,14 @@ class DALIPreprocessor(object):
             self.images, self.labels = daliop(
                 pipeline=pipe,
                 shapes=[(batch_size, height, width, 3), (batch_size, 1)],
-                dtypes=[tf.float32, tf.int32],
+                dtypes=[tf.float32, tf.float32],
                 device_id=device_id)
 
     def get_device_minibatches(self):
         with tf.device("/gpu:0"):
             self.labels -= 1 # Change to 0-based (don't use background class)
             #self.labels = tf.squeeze(self.labels)
-        return self.images, self.labels
+        return self.images, tf.cast(self.labels, tf.int64)
 
 
 def input_fn(is_training, data_dir, batch_size, num_epochs=1, num_gpus=None,
